modules/api/requirements.txt
# Bielik MVP API - Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.9
requests==2.31.0
pydantic==2.6.1
python-multipart==0.0.9
